pyproject.toml
[tool.poetry]
name = "collector"
version = "0.1.0"
description = ""
authors = ["apanchenko <apanchenko@quantumbrains.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.3"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

